refactor(web): reubica actions del workspace y elimina el dir coordinacion muerto
