+++ apps/web/src/components/organisms/expired-need-card.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { useActionState } from 'react';
-import { renewNeed } from '@/app/e/[slug]/coordinacion/actions';
+import { renewNeed } from '@/app/emergencies/[slug]/manage/actions';
 import type { components } from '@reliefhub/api-client';
-import type { ActionResult } from '@/app/e/[slug]/coordinacion/actions';
+import type { ActionResult } from '@/app/emergencies/[slug]/manage/actions';
 import { Button } from '@/components/atoms/button';
 import { ErrorMessage } from '@/components/atoms/error-message';
 import { useLocale } from '@/i18n/locale-context';
