+++ apps/web/src/components/organisms/need-detail.tsx
@@ -5,9 +5,9 @@ import {
   validateNeed,
   editNeed,
   discardNeed,
-} from '@/app/e/[slug]/coordinacion/actions';
+} from '@/app/emergencies/[slug]/manage/actions';
 import type { components } from '@reliefhub/api-client';
-import type { ActionResult } from '@/app/e/[slug]/coordinacion/actions';
+import type { ActionResult } from '@/app/emergencies/[slug]/manage/actions';
 import { Badge } from '@/components/atoms/badge';
 import { Button } from '@/components/atoms/button';
 import { ErrorMessage } from '@/components/atoms/error-message';
